Add InetAddress for Native
