+++ library/runtime-core/src/nativeMain/kotlin/io/matthewnelson/kmp/tor/runtime/core/internal/NativePlatform.kt
@@ -15,6 +15,9 @@
  **/
 package io.matthewnelson.kmp.tor.runtime.core.internal
 
+import io.matthewnelson.kmp.tor.core.api.annotation.InternalKmpTorApi
+import io.matthewnelson.kmp.tor.runtime.core.address.Port
+import org.kotlincrypto.endians.BigEndian.Companion.toBigEndian
 import kotlin.experimental.ExperimentalNativeApi
 
 internal actual val IsAndroidHost: Boolean get() {
@@ -33,4 +36,19 @@ internal actual val IsDarwinMobile: Boolean get() {
         OsFamily.WATCHOS-> true
         else -> false
     }
+}
+
+@InternalKmpTorApi
+public inline fun Port.toSinPort(): UShort = value.toSinPort()
+
+@PublishedApi
+internal fun Int.toSinPort(): UShort {
+    @OptIn(ExperimentalNativeApi::class)
+    if (!(Platform.isLittleEndian)) return toUShort()
+
+    return toShort()
+        .toBigEndian()
+        .toLittleEndian()
+        .toShort()
+        .toUShort()
 }
